requirements.txt
python-telegram-bot>=20.0
requests>=2.31.0
numpy>=1.26.0
pandas>=2.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
pytz>=2024.1
APScheduler==3.10.4

